+++ internal/models/productORM.go
@@ -4,13 +4,13 @@ type Product struct {
 	ID          string  `json:"product_id" gorm:"primaryKey"`
 	Name        string  `json:"name"`
 	Description string  `json:"description,omitempty"`
-	Price       float64 `json:"price"`
+	Price       float32 `json:"price"`
 	Quantity    int     `json:"quantity"`
 }
 
 type UpdateProduct struct {
 	Name        *string  `json:"name,omitempty"`
 	Description *string  `json:"description,omitempty"`
-	Price       *float64 `json:"price,omitempty"`
+	Price       *float32 `json:"price,omitempty"`
 	Quantity    *int     `json:"quantity,omitempty"`
 }
